first work brush

+++ src/engine/core/tool.d
@@ -10,6 +10,6 @@ interface Tool: SettingObject
     const(ImageReadAccess) cursor() const;
     void activate();
     void deactivate();
-    void mouse_eh( in ivec2 mpos, in DiMouseEvent me );
-    void keyboard_eh( in ivec2 mpos, in DiKeyboardEvent ke );
+    void mouse_eh( in vec2 mpos, in DiMouseEvent me );
+    void keyboard_eh( in vec2 mpos, in DiKeyboardEvent ke );
 }
